requirements.txt
flask>=3.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
requests>=2.31.0
